Server working
merge complete

+++ includes/webserv.hpp
@@ -52,7 +52,7 @@ struct ServerConfig
     std::vector<std::pair<std::string, int> >   ips_and_ports;
     std::vector<std::string>                    server_names;
     std::map<int, std::string>                  error_pages;
-    size_t                                      client_max_body_size;
+    unsigned long                               client_max_body_size;
     std::vector<LocationConfig>                 locations;
 };
